refactor: move handler functions
 Define the following functions and remove it from
MainWindowMenuHandler class

- MainWindowUI.createSearchWindow
- MainWindowUI.getTrimmedSelectedTextInMainWindow
- ProjectUICommands.projectExit
- ProjectUICommands.projectRestart
- ProjectUICommands.prepareForExit
- ProjectUICommands.openFile
- ProjectUICommands.openWriteableGlossaryFile
- ProjectUICommands.areTagsValid
