+++ src/org/omegat/gui/main/MainWindowUI.java
@@ -31,6 +31,7 @@
 
 package org.omegat.gui.main;
 
+import java.awt.Component;
 import java.awt.GraphicsEnvironment;
 import java.awt.Rectangle;
 import java.awt.event.WindowAdapter;
@@ -43,16 +44,20 @@ import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.swing.text.JTextComponent;
+
 import org.omegat.core.Core;
 import org.omegat.core.CoreEvents;
 import org.omegat.core.events.IApplicationEventListener;
 import org.omegat.core.events.IProjectEventListener;
 import org.omegat.core.search.SearchMode;
+import org.omegat.gui.editor.EditorUtils;
 import org.omegat.gui.search.SearchWindowController;
 import org.omegat.util.Log;
 import org.omegat.util.OConsts;
 import org.omegat.util.OStrings;
 import org.omegat.util.StaticUtils;
+import org.omegat.util.StringUtil;
 import org.omegat.util.gui.UIDesignManager;
 
 import com.vlsolutions.swing.docking.DockingDesktop;
@@ -110,6 +115,10 @@ public final class MainWindowUI {
         PerProjectLayoutHandler handler = new PerProjectLayoutHandler(mainWindow);
         CoreEvents.registerProjectChangeListener(handler);
         CoreEvents.registerApplicationEventListener(handler);
+    }
+
+    static void createSearchWindow(SearchMode mode) {
+        createSearchWindow(mode, getTrimmedSelectedTextInMainWindow());
     }
 
     static void createSearchWindow(SearchMode mode, String query) {
@@ -252,8 +261,7 @@ public final class MainWindowUI {
 
         // Ensure any "closed" Dockables are visible. These can be newly added
         // panes not included in an older layout file, or e.g. panes installed
-        // by
-        // plugins.
+        // by plugins.
         UIDesignManager.ensureDockablesVisible(mainWindow.desktop);
     }
 
@@ -312,5 +320,18 @@ public final class MainWindowUI {
         } catch (Exception e) {
             Log.log(e);
         }
+    }
+
+    public static String getTrimmedSelectedTextInMainWindow() {
+        String selection = null;
+        Component component = Core.getMainWindow().getApplicationFrame().getMostRecentFocusOwner();
+        if (component instanceof JTextComponent) {
+            selection = ((JTextComponent) component).getSelectedText();
+            if (!StringUtil.isEmpty(selection)) {
+                selection = EditorUtils.removeDirectionChars(selection);
+                selection = selection.trim();
+            }
+        }
+        return selection;
     }
 }
